feat(machine): wire provisioner worker with machine cloud instance set
This patch introduces the machine service into the provisioner worker,
via the service factory. Using the newly implemented machine cloud
instance (sub-) domain we are able to set the machine cloud instance for
the provisioned machines, previously known as instance data.

It must be noted that the current call to SetInstanceInfo through the
facades is maintained to keep the double-write working until we fully
migrate the machine domain to dqlite.