feat: provide a way to break out from interactive credentials prompt
This patch adds a way to stop prompting credential attributes *if* the
attributed is marked as being the last one and its value is non empty.
This was thought as a cheap(er) way of introducing a conditional exit of
the list of prompt steps, where we cannot do branching.
Also, a new enter function is created for the pollster lib, where we
provide a suffix in the case of an attribute being the last (breaking)
one, this way the user is aware of the mechanism.