fix: lxd VM instances correctly deploy now
There was a bug which prevented us deploying LXD VM instances on main.
The issue was that when trying to add a machine with constraints
"virt-type=virtual-machine" the machine never reached an active state
and the logs showed LXD errors complaining that the alias already
existed.

The actual problem was that on https://github.com/juju/juju/pull/17403
we added two aliases to the newly copied (to the local store) LXD image:
the juju constructed one (juju/ubuntu@24.04/amd64/vm) as well as
ubuntu@24.04 which is the original one. Removing this original alias
fixed the first "alias already exists" error.

The second error is that we incorrectly asumed that the image
fingerprint needed to be equal to the image alias, and therefore we set
the fingerprint to the original alias (which is ubuntu@24.04 - the same
for container and vms), therefore we could never start the VM instances
because the VM image wasn't there.
This is actually not needed, LXD will correctly keep the auto_update
field when copying the image to the local store even if we leave the
image fingerprint.
This was copied from lxd codebase
(https://github.com/canonical/lxd/blob/f2dded07c9c4c046d3484f50c8248171b65352b6/lxc/image.go#L255-L256)
but the difference is that they are only doing that in the case of
public images, but once we copy it to the local store the images are no
longer public.

As a side quest, the type (container or VM) field was added to the CopyImage parameters, this is not used but it will allow us easier debugging and correct images.