fix: replicaset update after removing a primary controller in HA
Currently in HA controllers, if the machine corresponding to the mongodb
PRIMARY goes down, then the replica set is never correctly updated and
therefore the resources are never destroyed correctly.

According to my analysis, this happens because the controller machine
that performs the StepDownPrimary does refresh the mongo session, and
therefore the session will contain the new conn sockets with the correct
addresses, but not the other machines, specially the new PRIMARY.
The resulting error is that when the new PRIMARY tries to update the
replica set then this mongo error is returned:
```
New config is rejected :: caused by :: replSetReconfig should only be run on a writable PRIMARY. Current state SECONDARY
```

The fix is therefore to refresh the session not only after the primary
step down (which will refresh the session for the machine going down)
but also for the other machines, just before trying to update the
replica set.