fix: race in application scale update
Currently we have a race in the `ChangeApplicationScale()` method of the
application service layer.
The issue is that we retrieve the current scale value and update it in
two separate transactions.
Even though we agreed not to have bussiness logic in the state layer,
doing this check under the same transaction is a necessity, so the
validation (for when the new desired scale is less than 0) is therefore
done in the state layer.

Tests were updated accordingly.