chore: replace the 0 constant by a real UUID v7 for alpha
The alpha space currently has a hard-coded ID `0`, which is inconvenient
both because the DDL has `uuid` in the space table (which `0` isn't) and
also because when we list the spaces on the CLI we get inconsistent
output between alpha and the manually created spaces:
```
Name   Space ID                              Subnets
alpha  0                                     172.31.0.0/20
beta   0195942d-54fc-77b6-99be-1c36e21da0d6  172.31.16.0/20
gamma  0195942d-9b23-7a6b-ba13-04bab64c628e  172.31.32.0/20
```

This patch addresses these issues and assigns a static constant UUID v7
to the alpha space.
The DDL had to be updated to take this new value into account, as well
as the tests.