chore: create DesiredReplicas for caas and move deployment type to core
This patch adds a new method to the kubernetes client which is actually
a subset of the returned structs by Application().
Also, since the goal of this is to have `DesiredReplicas()` as part of
the environs interface, we move the input type `DeploymentType` to core,
since leaving it in caas would create a circular dependency.